+++ youlai-mall-service-api/yshop-pms-api/src/main/java/com/fly4j/yshop/pms/feign/admin/factory/PmsFeignFallbackFactory.java
@@ -13,7 +13,7 @@ public class PmsFeignFallbackFactory implements FallbackFactory<PmsFeign> {
     public PmsFeign create(Throwable throwable) {
         return new PmsFeign() {
             @Override
-            public R getSpuById(Long id) {
+            public R getSpuById(Integer id) {
                 return null;
             }
